Moved HilbertMapper Class to a different file
